refactoring native-task-executor into native-task-handler
