+++ tee-worker/omni-executor/native-task-handler/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "native-call-executor"
+name = "native-task-handler"
 version = "0.1.0"
 authors = ['Trust Computing GmbH <info@litentry.com>']
 edition.workspace = true
